Implemented remote cursor displays on initial document connect
Needs testing on frontend to see if remote cursor positions are correct,
and there may be a bug where if someone is highlighting as someone connects,
the highlight region isn't showing.

Backend's new global data structure needs to be tested for disconnects and
if it's maintaining/deleting the connections properly or not.

Backend code needs heavy refactoring.